+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - PIM-9659: Asset Manager - Fix missing warning message when changing page via the breadcrumb with unsaved changes.
 - PIM-9676: Reference Entities - Fix missing warning message when changing page via the breadcrumb with unsaved changes.
 - PIM-9668: Asset attribute - Fix Text area + Rich text editor modes turn text attributes into infinite extendable fields when not using spaces
+- PIM-9664: Display Ziggy as asset image when the preview cannot be generated
 
 ## Improvements